# cache proxy configuration to get 12.5% speedup