@nateprewitt 2023-02-22
Hi @SonOfLilit,

Thanks for the suggestion! I'm not sure I'm seeing conclusive data for the improvements claim though. Requests has a fairly diverse user population. There are several variables we'd need to weigh including frequency of underlying configuration changes, number of proxies/configurations used and operating system before we can make any conclusions on performance.

There may be some minor performance improvements to be had, but introduction of a cache also introduces a new set of problems. Any changes in the underlying OS configurations may be ignored and there's no reasonable way to force cache eviction with this approach. We're also largely under a feature freeze, so I think it's unlikely we'll accept this at this time.