@SonOfLilit 2023-02-22
Thanks for the quick response.

Eviction is not reasonable to support here, but I don't see any reason to - would developers expect an existing session to react to changes in proxy bypass OS configuration? I'd expect to need to restart the process in this case, not just the session.

The performance claims are indeed Mac-only at this time. If I try this on Windows and Linux and see consistent >10% gains, will it make a difference?

Also, need help maintaining this project? It's my favorite python package since the day it came out, I can find some time in my day to do some code reviews.