@nateprewitt 2023-02-22
> Eviction is not reasonable to support here, but I don't see any reason to [...] I'd expect to need to restart the process in this case, not just the session.

So that's an assumption, but it's not a data point without validation. The first concerns would be resolution from the[ Windows Registry](https://github.com/psf/requests/blob/main/requests/utils.py#L68-L116) and macOS system configurations via [SCDynamicCopyStoreProxies](https://developer.apple.com/documentation/systemconfiguration/1517088-scdynamicstorecopyproxies) ([cpython ref](https://github.com/python/cpython/blob/main/Modules/_scproxy.c#L171-L216)). I'm not claiming these are dynamic during a process' lifecycle because I haven't tested it. It would not surprise me if both are queried in real time and subject to change.

> The performance claims are indeed Mac-only at this time.

The data currently is for a fairly naive idea of how proxy configurations work. They demonstrate a single use-case experiences improvement at the potential cost of incorrect data. We'll always optimize for correctness over speed.

There's likely an eventuality where we could work through this to reach a conclusion but I don't think it's worth the time trade-off to do it right.

> Also, need help maintaining this project?

We appreciate the offer but I don't think code reviewing is the current bottleneck. All changes need to be weighed for potential impact and added complexity to the library. For the time being, we're relatively OK with the libraries current functionality and don't intend to make any significant changes.