@SonOfLilit 2023-02-23
On Thu, Feb 23, 2023, 01:20 Nate Prewitt ***@***.***> wrote:

> Eviction is not reasonable to support here, but I don't see any reason to
> [...] I'd expect to need to restart the process in this case, not just the
> session.
>
> So that's an assumption, but it's not a data point without validation.
> [..] I'm not claiming these are dynamic during a process' lifecycle because
> I haven't tested it. It would not surprise me if both are queried in real
> time and subject to change.
>
Oh no, I'm almost sure they're queried in real time and subject to change.
The assumption is about user expectations. I'm essentially saying "the
current behavior is probably perceived as a bit more buggy than the
proposed change". I know this is an always-wrong thing to say, but I'd be
surprised if any users ever considered what behavior they want in this
case, other than an obligatory xkcd.com/1172 guy of course.

We appreciate the offer but I don't think code reviewing is the current
> bottleneck. All changes need to be weighed for potential impact and added
> complexity to the library.
>
Yeah, that's what I meant by code review for a library that the entire
software world depends on.

Alright, I'll stop taking your time, thanks for doing all this!

> —
> Reply to this email directly, view it on GitHub
> <https://github.com/psf/requests/pull/6360#issuecomment-1440989131>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAADU7NMXZNNR6WSALOSF5LWY2NMXANCNFSM6AAAAAAVE4DRTQ>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>